Changed terraform state file bucket names (HW-22)
Changed due to google bucket name considerations "Bucket names reside
in a single Cloud Storage namespace, which means that every bucket name
must be unique. If you try to create a bucket with a name that is
already assigned to an existing bucket, Cloud Storage responds with
an error message. However, once you delete a bucket, its name can
be reused by you or another user when creating a new bucket."